# api/swagger.yaml
definitions:
...
  app.ModelDTO:
    properties:
      created_at:
        type: integer
      desc:
        type: string
      download_count:
        type: integer
      fullname:
        type: string
      id:
        type: string
      labels:
        $ref: '#/definitions/app.ModelLabelsDTO'
      like_count:
        type: integer
      name:
        type: string
      owner:
        type: string
      updated_at:
        type: integer
      visibility:
        type: string
    type: object
  app.ModelLabelsDTO:
    properties:
      frameworks:
        items:
          type: string
        type: array
      license:
        type: string
      others:
        items:
          type: string
        type: array
      task:
        type: string
    type: object
  app.ModelsDTO:
    properties:
      models:
        items:
          $ref: '#/definitions/repository.ModelSummary'
        type: array
      total:
        type: integer
    type: object
...
  controller.ResponseData:
    properties:
      code:
        type: string
      data: {}
      msg:
        type: string
    type: object
...
  controller.modelDetail:
    properties:
      avatar_id:
        type: string
      created_at:
        type: integer
      desc:
        type: string
      download_count:
        type: integer
      fullname:
        type: string
      id:
        type: string
      labels:
        $ref: '#/definitions/app.ModelLabelsDTO'
      like_count:
        type: integer
      liked:
        type: boolean
      name:
        type: string
      owner:
        type: string
      updated_at:
        type: integer
      visibility:
        type: string
    type: object
  controller.modelInfo:
    properties:
      avatar_id:
        type: string
      desc:
        type: string
      download_count:
        type: integer
      fullname:
        type: string
      id:
        type: string
      like_count:
        type: integer
      name:
        type: string
      owner:
        type: string
      task:
        type: string
      updated_at:
        type: integer
    type: object
  controller.modelsInfo:
    properties:
      models:
        items:
          $ref: '#/definitions/controller.modelInfo'
        type: array
      total:
        type: integer
    type: object
...
  controller.reqToCreateModel:
    properties:
      desc:
        type: string
      fullname:
        type: string
      init_readme:
        type: boolean
      license:
        type: string
      name:
        type: string
      owner:
        type: string
      visibility:
        type: string
    type: object
...
  controller.reqToUpdateModel:
    properties:
      desc:
        type: string
      fullname:
        type: string
      name:
        type: string
      visibility:
        type: string
    type: object
...
  controller.userModelsInfo:
    properties:
      avatar_id:
        type: string
      models:
        items:
          $ref: '#/definitions/repository.ModelSummary'
        type: array
      owner:
        type: string
      total:
        type: integer
    type: object
...
  repository.ModelSummary:
    properties:
      desc:
        type: string
      download_count:
        type: integer
      fullname:
        type: string
      id:
        type: string
      like_count:
        type: integer
      name:
        type: string
      owner:
        type: string
      task:
        type: string
      updated_at:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /v1/model:
    get:
      consumes:
      - application/json
      description: list global public model
      parameters:
      - description: name of model
        in: query
        name: name
        type: string
      - description: task label
        in: query
        name: task
        type: string
      - description: other labels, separate multiple each ones with commas
        in: query
        name: others
        type: string
      - description: license label
        in: query
        name: license
        type: string
      - description: framework labels, separate multiple each ones with commas
        in: query
        name: frameworks
        type: string
      - description: whether to calculate the total
        in: query
        name: count
        type: boolean
      - description: 'sort types: most_likes, alphabetical, most_downloads, recently_updated,
          recently_created'
        in: query
        name: sort_by
        type: string
      - description: page num which starts from 1
        in: query
        name: page_num
        type: integer
      - description: count per page
        in: query
        name: count_per_page
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.modelsInfo'
      summary: ListGlobal
      tags:
      - ModelWeb
    post:
      consumes:
      - application/json
      description: create model
      parameters:
      - description: body of creating model
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.reqToCreateModel'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: Create
      tags:
      - Model
  /v1/model/{id}:
    delete:
      consumes:
      - application/json
      description: delete model
      parameters:
      - description: id of model
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete
      tags:
      - Model
    put:
      consumes:
      - application/json
      description: update model
      parameters:
      - description: id of model
        in: path
        name: id
        required: true
        type: string
      - description: body of updating model
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.reqToUpdateModel'
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: Update
      tags:
      - Model
  /v1/model/{owner}:
    get:
      consumes:
      - application/json
      description: list model
      parameters:
      - description: owner of model
        in: path
        name: owner
        required: true
        type: string
      - description: name of model
        in: query
        name: name
        type: string
      - description: whether to calculate the total
        in: query
        name: count
        type: boolean
      - description: 'sort types: most_likes, alphabetical, most_downloads, recently_updated,
          recently_created'
        in: query
        name: sort_by
        type: string
      - description: page num which starts from 1
        in: query
        name: page_num
        type: integer
      - description: count per page
        in: query
        name: count_per_page
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.userModelsInfo'
      summary: List
      tags:
      - ModelWeb
  /v1/model/{owner}/{name}:
    get:
      consumes:
      - application/json
      description: get model
      parameters:
      - description: owner of model
        in: path
        name: owner
        required: true
        type: string
      - description: name of model
        in: path
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.modelDetail'
      summary: Get
      tags:
      - ModelWeb
...
  /v2/model:
    get:
      consumes:
      - application/json
      description: list global public model
      parameters:
      - description: name of model
        in: query
        name: name
        type: string
      - description: task label
        in: query
        name: task
        type: string
      - description: owner of model
        in: query
        name: owner
        required: true
        type: string
      - description: other labels, separate multiple each ones with commas
        in: query
        name: others
        type: string
      - description: license label
        in: query
        name: license
        type: string
      - description: framework labels, separate multiple each ones with commas
        in: query
        name: frameworks
        type: string
      - description: whether to calculate the total
        in: query
        name: count
        type: boolean
      - description: 'sort types: most_likes, alphabetical, most_downloads, recently_updated,
          recently_created'
        in: query
        name: sort_by
        type: string
      - description: page num which starts from 1
        in: query
        name: page_num
        type: integer
      - description: count per page
        in: query
        name: count_per_page
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.ModelsDTO'
      summary: List
      tags:
      - ModelRestful
  /v2/model/{owner}/{name}:
    get:
      consumes:
      - application/json
      description: get model
      parameters:
      - description: owner of model
        in: path
        name: owner
        required: true
        type: string
      - description: name of model
        in: path
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.ModelDTO'
      summary: Get
      tags:
      - ModelRestful
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and api Bearer.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"